requirements.txt
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.0
deap==1.4.1
filelock==3.12.4
fonttools==4.43.0
fsspec==2023.9.2
graphviz==0.20.1
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
logparser3==1.0.4
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.24.4
packaging==23.2
pandas==2.0.3
Pillow==10.0.1
pydot==1.4.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2022.3.2
scikit-learn==1.3.1
scipy==1.10.1
six==1.16.0
sklearn==0.0.post9
sympy==1.12
threadpoolctl==3.2.0
torch==2.1.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
zipp==3.17.0
